Extract announcement form in own component.  New common html structure in AnnouncementCard with reworked conditions.  Make announcement dismiss as optional.
